Remove hardcoded readgroup filter.
`samtools view -r readgroup` outputs read only if the readgroup is
really set to `readgroup` or if no readgroups are set. In the future we
could make this configurable, but on the other hand I don't think
galaxy can really benefit from this. Instead of a multibam file with
multiple readgroups having multiple bam files allows for easier
parallelization.